Allows bot to connect if initially started while at slash screen
If you were to attempt to start the bot while the switch is currently at the splash screen it will attempt to connect but never be able to. 

Added a single instance of pressing A once - Tested this with attemptCount == 2 and it connects just in time for attempt 5, when ``attemptCount == 1`` is the conditional it works as intended,

You can now start the bot regardless if you are at the splash screen, or actually in the overworld, and it will eventually connect, update the seed, and run the raid - I added this because I personally always start it when its on the splash screen and got annoyed having to get up to press a and start again.